Fixes to DTI pipeline: 1) do not remove reorient file when conventions match 2) fix for allowing custom gradient table and 3) re-generate pipeline_pb2.py with latest protoc (this maybe does not need to be checked in since it is generated during setup.py?)

+++ cmp/util.py
@@ -105,7 +105,10 @@ def reorient(src, ref, log):
     shutil.move(tmp2, src)
     log.info("File %s written" % src)
     
-    os.remove(tmpsrc)
+    # Only remove the temporary file if the conventions did not match.  Otherwise,
+    # we end up removing the output.
+    if tmpsrc != src:
+        os.remove(tmpsrc)
     log.info("Remove temporary file %s" % tmpsrc)   
 
 def DTB_viewer():
